ftp://ftp.krokus.ru/pub/OpenBSD/distfiles/xscreensaver-4.22.tar.gz -rw-r--r-- 1 zblaxell zblaxell 4620066 Sep  4  2005 xscreensaver-4.22.tar.gz f40db41ef221f7ed42c39e77007ce05b4f846306  xscreensaver-4.22.tar.gz

+++ hacks/rd-bomb.c
@@ -198,7 +198,8 @@ pixack_frame(char *pix_buf)
     int ii = i + 1;
     char *q = pix_buf + width * i;
     short *qq = ((short *) pix_buf) + width * i;
-    long  *qqq = ((long *) pix_buf) + width * i;
+/*  long  *qqq = ((long *) pix_buf) + width * i;  -- crashes on Alpha */
+    int   *qqq = ((int  *) pix_buf) + width * i;
     ushort *i1 = r1 + 1 + w2 * ii;
     ushort *i2 = r2 + 1 + w2 * ii;
     ushort *o1 = r1b + 1 + w2 * ii;
